Create configuration.h for SKR Mini E3 v3
Create configuration.h for SKR Mini E3 v3 for Ender 3 V2 Repo.
This Configuration.h File is migrated from the Ender 3 Pro Repo.
Including:
Filament Runout Sensor activated
CR / BL Touch Activated  (UBL 10*10 Mesh)
Probe 5 on 2 times
M37 for calculating remaining Time

It is my first time i changed Firmware parameters. the Firmware is in use on my Ender 3 v2, but is seen as under development. It prints, but some things are not finished (not every icon is visible on LCD,...).
Also i promote, that a Picture of the Cable mod is added in here for beter Understanding. I can provide this.